formatting

+++ packages/hardhat-ts/contracts/GEONFT.sol
@@ -25,7 +25,11 @@ contract GEONFT is ERC721, ERC721Enumerable, ERC721URIStorage, ERC721Burnable, O
         return "";
     }
 
-    function safeMint(address to, string memory uri, string memory geoJson) public {
+    function safeMint(
+        address to,
+        string memory uri,
+        string memory geoJson
+    ) public {
         uint256 tokenId = _tokenIdCounter.current();
         _tokenIdCounter.increment();
         _safeMint(to, tokenId);
@@ -42,47 +46,67 @@ contract GEONFT is ERC721, ERC721Enumerable, ERC721URIStorage, ERC721Burnable, O
     }
 
     // Returns an array of tokenIds
-    function getAllTokens() public view returns (uint[] memory, string[] memory, string[] memory) {
-        uint totalTokens = totalSupply();
-        uint[] memory _tokenIds = new uint[](totalTokens);
+    function getAllTokens()
+        public
+        view
+        returns (
+            uint256[] memory,
+            string[] memory,
+            string[] memory
+        )
+    {
+        uint256 totalTokens = totalSupply();
+        uint256[] memory _tokenIds = new uint256[](totalTokens);
         string[] memory _uris = new string[](totalTokens);
         string[] memory _geoJsons = new string[](totalTokens);
-        uint i;
+        uint256 i;
 
-        for (i=0;i<totalTokens;i++){
+        for (i = 0; i < totalTokens; i++) {
             _tokenIds[i] = ERC721Enumerable.tokenByIndex(i);
             _uris[i] = tokenURI(_tokenIds[i]);
             _geoJsons[i] = geoJsons[_tokenIds[i]];
         }
-        return (_tokenIds, _uris, _geoJsons );
+        return (_tokenIds, _uris, _geoJsons);
     }
 
     // Returns an array of tokenIds, URIs for an owner address
-    function getTokensByOwner(address owner) public view returns (uint[] memory, string[] memory, string[] memory) {
-        uint totalTokensForOwner = ERC721.balanceOf(owner);
-        uint[] memory _tokenIds = new uint[](totalTokensForOwner);
+    function getTokensByOwner(address owner)
+        public
+        view
+        returns (
+            uint256[] memory,
+            string[] memory,
+            string[] memory
+        )
+    {
+        uint256 totalTokensForOwner = ERC721.balanceOf(owner);
+        uint256[] memory _tokenIds = new uint256[](totalTokensForOwner);
         string[] memory _uris = new string[](totalTokensForOwner);
         string[] memory _geoJsons = new string[](totalTokensForOwner);
-        uint i;
+        uint256 i;
 
-        for (i=0;i<totalTokensForOwner;i++){
+        for (i = 0; i < totalTokensForOwner; i++) {
             _tokenIds[i] = ERC721Enumerable.tokenOfOwnerByIndex(owner, i);
             _uris[i] = tokenURI(_tokenIds[i]);
             _geoJsons[i] = geoJsons[_tokenIds[i]];
         }
-        return (_tokenIds, _uris, _geoJsons );
+        return (_tokenIds, _uris, _geoJsons);
     }
 
     // The following functions are overrides required by Solidity.
 
-    function _beforeTokenTransfer(address from, address to, uint256 tokenId)
-        internal
-        override(ERC721, ERC721Enumerable)
-    {
+    function _beforeTokenTransfer(
+        address from,
+        address to,
+        uint256 tokenId
+    ) internal override(ERC721, ERC721Enumerable) {
         super._beforeTokenTransfer(from, to, tokenId);
     }
 
-    function _burn(uint256 tokenId) internal override(ERC721, ERC721URIStorage) {
+    function _burn(uint256 tokenId)
+        internal
+        override(ERC721, ERC721URIStorage)
+    {
         super._burn(tokenId);
     }
